requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg2==1.19.4
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
